Method 2 for context method in the PushChat
